backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.9
pandas==2.2.3
openpyxl==3.1.5
supabase==2.15.0
python-dotenv==1.0.1
thefuzz==0.22.1
python-Levenshtein==0.25.1
